+++ circle.py
@@ -1,5 +1,4 @@
 import math
-import unittest
 
 
 def area(r: float) -> float:
@@ -18,21 +17,3 @@ def perimeter(r: float) -> float:
     :returns perimeter
     """
     return 2 * math.pi * r
-
-
-class CircleTestCase(unittest.TestCase):
-    def test_zero_area(self):
-        res = area(0)
-        self.assertEqual(res, 0)
-
-    def test_area(self):
-        res = area(10)
-        self.assertEqual(res, 100 * math.pi)
-
-    def test_zero_perimeter(self):
-        res = perimeter(0)
-        self.assertEqual(res, 0)
-
-    def test_perimeter(self):
-        res = perimeter(10)
-        self.assertEqual(res, 20 * math.pi)
