+++ square.py
@@ -1,6 +1,3 @@
-import unittest
-
-
 def area(a: float) -> float:
     """
     Calculates the area of the square
@@ -17,21 +14,3 @@ def perimeter(a: float) -> float:
     :returns perimeter
     """
     return 4 * a
-
-
-class SquareTestCase(unittest.TestCase):
-    def test_zero_area(self):
-        res = area(0)
-        self.assertEqual(res, 0)
-
-    def test_area(self):
-        res = area(10)
-        self.assertEqual(res, 100)
-
-    def test_zero_perimeter(self):
-        res = perimeter(0)
-        self.assertEqual(res, 0)
-
-    def test_perimeter(self):
-        res = perimeter(10)
-        self.assertEqual(res, 40)
