+++ test/test_circle.py
@@ -1,0 +1,22 @@
+import math
+import unittest
+
+from circle import area, perimeter
+
+
+class CircleTestCase(unittest.TestCase):
+    def test_zero_area(self):
+        res = area(0)
+        self.assertEqual(res, 0)
+
+    def test_area(self):
+        res = area(10)
+        self.assertEqual(res, 100 * math.pi)
+
+    def test_zero_perimeter(self):
+        res = perimeter(0)
+        self.assertEqual(res, 0)
+
+    def test_perimeter(self):
+        res = perimeter(10)
+        self.assertEqual(res, 20 * math.pi)
