+++ test/test_rectangle.py
@@ -1,0 +1,21 @@
+import unittest
+
+from rectangle import area, perimeter
+
+
+class RectangleTestCase(unittest.TestCase):
+    def test_zero_mul(self):
+        res = area(10, 0)
+        self.assertEqual(res, 0)
+
+    def test_square_mul(self):
+        res = area(10, 10)
+        self.assertEqual(res, 100)
+
+    def test_zero_perimeter(self):
+        res = perimeter(0, 0)
+        self.assertEqual(res, 0)
+
+    def test_perimeter(self):
+        res = perimeter(10, 10)
+        self.assertEqual(res, 40)
