+++ test/test_square.py
@@ -1,0 +1,21 @@
+import unittest
+
+from square import area, perimeter
+
+
+class SquareTestCase(unittest.TestCase):
+    def test_zero_area(self):
+        res = area(0)
+        self.assertEqual(res, 0)
+
+    def test_area(self):
+        res = area(10)
+        self.assertEqual(res, 100)
+
+    def test_zero_perimeter(self):
+        res = perimeter(0)
+        self.assertEqual(res, 0)
+
+    def test_perimeter(self):
+        res = perimeter(10)
+        self.assertEqual(res, 40)
